converting to mongodb friendly java collections and removing maps with dotted keys

+++ src/main/scala/ai/diffy/ApiController.scala
@@ -134,7 +134,7 @@ class ApiController(
     if(id.isEmpty) {
       RequestPurgedException
     } else {
-      repository.findById(id.toLong) map { case dr =>
+      repository.findById(id) map { case dr =>
         Renderer.differenceResult(
           dr,
           includeRequest
